possible fix for #23, support jshint identifier parser
This fixes #23 for me, however I am not merging this into
master until I see jshint updated again and these new
scripts work.

The longer-term fix for issues like this is to require that
jshint is installed locally as a node module and use it
instead of the jshint we are installing locally in the
bundle.